+++ plugins/falcon-slider/resources/views/frontend/render.blade.php
@@ -77,7 +77,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 @endif
 @if(!empty($fsFonts))
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?{{ collect($fsFonts)->map(fn($f) => 'family=' . str_replace(' ', '+', $f) . ':ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,400;1,700')->implode('&') }}&display=swap">
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?{{ collect($fsFonts)->map(fn($f) => 'family=' . str_replace(' ', '+', $f) . ':ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,400;1,700')->implode('&') }}&display=swap">
 @endif
 @if(!empty($fsIcons))
     {{-- Subset: only the icon glyphs this slider actually uses --}}